Extract DSL into separate modules
