+++ spec/compendium/dsl/metric_spec.rb
@@ -1,0 +1,57 @@
+require 'spec_helper'
+require 'compendium'
+require 'compendium/dsl'
+
+describe Compendium::DSL::Metric do
+  subject do
+    Class.new do
+      extend Compendium::DSL
+    end
+  end
+
+  describe '#metric' do
+    let(:metric_proc) { -> { :metric } }
+
+    before do
+      subject.query :test
+      subject.metric :test_metric, metric_proc, through: :test
+    end
+
+    it 'should add a metric to the given query' do
+      expect(subject.queries[:test].metrics.first.name).to eq(:test_metric)
+    end
+
+    it 'should set the metric command' do
+      expect(subject.queries[:test].metrics.first.command).to eq(metric_proc)
+    end
+
+    context 'when through is specified' do
+      it 'should raise an error if specified for an invalid query' do
+        expect { subject.metric :test_metric, metric_proc, through: :fake }.to raise_error ArgumentError, 'query fake is not defined'
+      end
+
+      it 'should allow metrics to be defined with a block' do
+        subject.metric :block_metric, through: :test do
+          123
+        end
+
+        expect(subject.queries[:test].metrics[:block_metric].run(self, nil)).to eq(123)
+      end
+
+      it 'should allow metrics to be defined with a lambda' do
+        subject.metric :block_metric, -> (*) { 123 }, through: :test
+        expect(subject.queries[:test].metrics[:block_metric].run(self, nil)).to eq(123)
+      end
+    end
+
+    context 'when through is not specified' do
+      before { subject.metric(:no_through_metric) { |data| data } }
+
+      specify { expect(subject.queries).to include :__metric_no_through_metric }
+
+      it 'should return the result of the query as the result of the metric' do
+        expect(subject.queries[:__metric_no_through_metric].metrics[:no_through_metric].run(self, [123])).to eq(123)
+      end
+    end
+  end
+end
